Package made more solid
BOM Stripping is done by default

The Writer no longer implements JsonSerializable and IteratorAggregate
The Writer can only writes new records and output the full CSV

Removed all deprecated methods from the Reader and the Writer

Added

- AbstractCsv::getHeader
- AbstractCsv::getHeaderOffset
- AbstractCsv::setHeader

Removed

- Reader::each
- Reader::fetch
- Reader::fetchAssoc
- Reader::fetchPairsWithoutDuplicates
